requirements.txt
instapy
instabot
pillow
